Restucture complete.
